+++ module-api/src/main/java/com/depromeet/inspiration/member/ResultCode.java
@@ -1,0 +1,4 @@
+package com.depromeet.inspiration.member;
+
+public class ResultCode {
+}
