+++ module-api/src/main/java/com/depromeet/inspiration/member/ResultResponse.java
@@ -1,0 +1,10 @@
+package com.depromeet.inspiration.member;
+
+public class ResultResponse {
+
+    private int status;
+    private String code;
+    private String message;
+    private Object data;
+
+}
