--- module-api/src/main/java/inspiration/member/ResultCode.java
@@ -1,4 +1,0 @@
-package inspiration.member;
-
-public class ResultCode {
-}
