--- module-api/src/main/java/inspiration/member/ResultResponse.java
@@ -1,10 +1,0 @@
-package inspiration.member;
-
-public class ResultResponse {
-
-    private int status;
-    private String code;
-    private String message;
-    private Object data;
-
-}
