+++ module-common/src/main/java/com/depromeet/inspiration/config/JpaAuditingConfig.java
@@ -1,0 +1,9 @@
+package com.depromeet.inspiration.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfig {
+}
