+++ module-common/src/main/java/com/depromeet/inspiration/domain/date/BaseTimeEntity.java
@@ -1,0 +1,22 @@
+package com.depromeet.inspiration.domain.date;
+
+import lombok.Getter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import javax.persistence.EntityListeners;
+import javax.persistence.MappedSuperclass;
+import java.time.LocalDateTime;
+
+@Getter
+@EntityListeners(value = {AuditingEntityListener.class})
+@MappedSuperclass
+public abstract class BaseTimeEntity {
+
+    @CreatedDate
+    private LocalDateTime createdDate;
+
+    @LastModifiedDate
+    private LocalDateTime updatedDate;
+}
