+++ module-common/src/main/java/com/depromeet/inspiration/enumeration/ExceptionType.java
@@ -1,0 +1,16 @@
+package com.depromeet.inspiration.enumeration;
+
+public enum ExceptionType {
+
+    POST_NOT_FOUND("존재하지 않는 리소스입니다.");
+
+    private final String message;
+
+    ExceptionType(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}
