+++ module-common/src/main/java/com/depromeet/inspiration/exception/PostNotFoundException.java
@@ -1,0 +1,14 @@
+package com.depromeet.inspiration.exception;
+
+import com.depromeet.inspiration.enumeration.ExceptionType;
+
+public class PostNotFoundException extends RuntimeException {
+
+    public PostNotFoundException() {
+        super(ExceptionType.POST_NOT_FOUND.getMessage());
+    }
+
+    public PostNotFoundException(String message) {
+        super(message);
+    }
+}
