--- module-common/src/main/java/inspiration/config/JpaAuditingConfig.java
@@ -1,9 +1,0 @@
-package inspiration.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {
-}
