--- module-common/src/main/java/inspiration/exception/PostNotFoundException.java
@@ -1,14 +1,0 @@
-package inspiration.exception;
-
-import inspiration.enumeration.ExceptionType;
-
-public class PostNotFoundException extends RuntimeException {
-
-    public PostNotFoundException() {
-        super(ExceptionType.POST_NOT_FOUND.getMessage());
-    }
-
-    public PostNotFoundException(String message) {
-        super(message);
-    }
-}
