+++ module-domain/src/main/java/com/depromeet/inspiration/member/MemberRepository.java
@@ -1,0 +1,6 @@
+package com.depromeet.inspiration.member;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+}
