--- module-domain/src/main/java/domain/member/MemberRepository.java
@@ -1,6 +1,0 @@
-package domain.member;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-}
