+++ module-web/src/main/java/com/depromeet/inspiration/v1/ResultCode.java
@@ -1,0 +1,4 @@
+package com.depromeet.inspiration.v1;
+
+public class ResultCode {
+}
