+++ module-web/src/main/java/com/depromeet/inspiration/v1/ResultResponse.java
@@ -1,0 +1,10 @@
+package com.depromeet.inspiration.v1;
+
+public class ResultResponse {
+
+    private int status;
+    private String code;
+    private String message;
+    private Object data;
+
+}
