+++ module-web/src/main/java/com/depromeet/inspiration/v1/advice/ErrorResponse.java
@@ -1,0 +1,16 @@
+package com.depromeet.inspiration.v1.advice;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+public class ErrorResponse {
+
+    private String message;
+
+    public static ErrorResponse of(String message) {
+        return new ErrorResponse(message);
+    }
+}
