+++ module-web/src/main/java/com/depromeet/inspiration/v1/member/MemberController.java
@@ -1,0 +1,7 @@
+package com.depromeet.inspiration.v1.member;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class MemberController {
+}
