--- module-web/src/main/java/inspiration/Application.java
@@ -1,13 +1,0 @@
-package inspiration;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
-
-@SpringBootApplication
-@EntityScan(basePackages = {"domain"})
-public class Application {
-    public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
-    }
-}
