--- module-web/src/main/java/inspiration/v1/ResultCode.java
@@ -1,4 +1,0 @@
-package inspiration.v1;
-
-public class ResultCode {
-}
