--- module-web/src/main/java/inspiration/v1/ResultResponse.java
@@ -1,10 +1,0 @@
-package inspiration.v1;
-
-public class ResultResponse {
-
-    private int status;
-    private String code;
-    private String message;
-    private Object data;
-
-}
