--- module-web/src/main/java/inspiration/v1/advice/ErrorResponse.java
@@ -1,16 +1,0 @@
-package inspiration.v1.advice;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class ErrorResponse {
-
-    private String message;
-
-    public static ErrorResponse of(String message) {
-        return new ErrorResponse(message);
-    }
-}
