--- module-web/src/main/java/inspiration/v1/member/MemberController.java
@@ -1,7 +1,0 @@
-package inspiration.v1.member;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class MemberController {
-}
